Move redis_vizsjon_cache to app/helpers #3344
