+++ app/helpers/redis_vizjson_cache.rb
