+++ app/models/user.rb
@@ -4,7 +4,7 @@ require_relative './user/user_decorator'
 require_relative './user/oauths'
 require_relative './synchronization/synchronization_oauth'
 require_relative './visualization/member'
-require_relative './visualization/redis_vizjson_cache'
+require_relative '../helpers/redis_vizjson_cache'
 require_relative './visualization/collection'
 require_relative './user/user_organization'
 require_relative './synchronization/collection.rb'
